@@ -1,7 +1,7 @@
 package net.yupol.transmissionremote.app;
 
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.List;
 
