@@ -1,6 +1,6 @@
 package net.yupol.transmissionremote.app;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 public abstract class BaseFragment extends Fragment implements OnBackPressedListener {
 
