@@ -1,6 +1,6 @@
 package net.yupol.transmissionremote.app.drawer;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.TextView;
 
