@@ -1,7 +1,7 @@
 package net.yupol.transmissionremote.app.drawer;
 
-import android.support.annotation.Nullable;
-import android.support.annotation.StringRes;
+import androidx.annotation.Nullable;
+import androidx.annotation.StringRes;
 
 import com.mikepenz.materialdrawer.model.PrimaryDrawerItem;
 
