@@ -1,6 +1,6 @@
 package net.yupol.transmissionremote.app.filtering;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import net.yupol.transmissionremote.app.R;
 import net.yupol.transmissionremote.app.model.json.Torrent;
