@@ -1,6 +1,6 @@
 package net.yupol.transmissionremote.app.model;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.mikepenz.fontawesome_typeface_library.FontAwesome;
 import com.mikepenz.iconics.typeface.IIcon;
