@@ -1,6 +1,6 @@
 package net.yupol.transmissionremote.app.model;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public enum TrackerState {
 
