@@ -2,7 +2,7 @@ package net.yupol.transmissionremote.app.model.json;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.google.api.client.util.Key;
 
