@@ -1,8 +1,8 @@
 package net.yupol.transmissionremote.app.notifications;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import com.evernote.android.job.Job;
