@@ -1,7 +1,7 @@
 package net.yupol.transmissionremote.app.opentorrent;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
