@@ -1,9 +1,9 @@
 package net.yupol.transmissionremote.app.preferences;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.MenuItem;
 
 public class AdvancedPreferencesActivity extends AppCompatActivity {
