@@ -3,7 +3,7 @@ package net.yupol.transmissionremote.app.torrentdetails;
 import android.app.Activity;
 import android.content.res.TypedArray;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.text.InputFilter;
 import android.util.AttributeSet;
 import android.util.Log;
