@@ -3,7 +3,7 @@ package net.yupol.transmissionremote.app.transport;
 import android.content.Context;
 import android.os.Handler;
 import android.os.Looper;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import com.octo.android.robospice.SpiceManager;
