@@ -2,8 +2,8 @@ package net.yupol.transmissionremote.app.utils;
 
 import android.content.Context;
 import android.content.res.ColorStateList;
-import android.support.annotation.ColorInt;
-import android.support.v4.content.ContextCompat;
+import androidx.annotation.ColorInt;
+import androidx.core.content.ContextCompat;
 import android.util.TypedValue;
 
 public class ColorUtils {
