@@ -1,6 +1,6 @@
 package net.yupol.transmissionremote.app.torrentdetails;
 
-import android.databinding.DataBindingUtil;
+import androidx.databinding.DataBindingUtil;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
